// src/app/models/individual/individualModelInterface.ts
export interface IndividdualModelInterface {
    PersonType: string;
    Names: string;
    LastName: string;
    SecondLastname: string;
    Birthdate: string;
    Street: string;
    Zipcode: string;
    State: string;
    City: string;
    Package: string;
    PaymentMethod: string;
    ValidFrom: string;
    ValidThrough: string;
    IssueCost: string;
    totalInsurees: string;
    CommissionPercentage: string;
    RFC: string;
    Email: string;
    Phone: string;
    Client_Id: number; // Asumí que Client_Id es un número, ajusta según tu necesidad
    AgentID: string;
    SubAgentID: string;
    Category: string;
}